move autocomplete into new completion package
